refactored faq page and route more questions button to contact-us page
