+++ frontend/src/Component/button/index.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import './button.scss'
 
-const Button = ({children, name, width, className}) => {
+const Button = ({children, name, width, className, onClick}) => {
   return (
       <button className={`btn ${className} `} style={{
         width: width
-    }}>{children} {name}</button>
+    }} onClick={onClick}>{children} {name}</button>
   )
 }
