+++ frontend/src/Style/App.scss
@@ -24,3 +24,7 @@ body {
   flex-direction: column;
   min-height: 100vh;
 }
+
+button {
+  cursor: pointer;
+}
