--- frontend/src/pages/FAQ/FAQcomp/Downloads.jsx
@@ -1,36 +1,0 @@
-import React, { useState } from "react";
-import Style from "../faq.module.scss";
-
-import { FaqItems } from "./FaqItems";
-
-import { DownlaodsItems } from "./FaqModule";
-
-export const Downloads = ({ header }) => {
-  const [clicked, setclicked] = useState("0");
-  const handleToggle = (index) => {
-    console.log(index);
-    if (clicked === index) {
-      return setclicked("0");
-    }
-    setclicked(index);
-  };
-
-  return (
-    <>
-      <section className={Style.faqContainer}>
-        <h2>{header}</h2>
-
-        {DownlaodsItems.map((faq, index) => {
-          return (
-            <FaqItems
-              key={index}
-              faq={faq}
-              onToggle={() => handleToggle(index)}
-              active={clicked === index}
-            />
-          );
-        })}
-      </section>
-    </>
-  );
-};
