+++ frontend/src/pages/FAQ/FaqModule.js
@@ -1,0 +1,130 @@
+export const GettingStartedItems = [
+  {
+    title: "What does Certawi Certificate Generator do?",
+    body: [
+      "It is a complete and easy to use website that can generate and send certificate to uploaded CSV lists.",
+    ],
+  },
+
+  {
+    title: "How do I sign up for Certawi Certificate Generator?",
+    body: ["You can sign up with your email account."],
+  },
+  {
+    title: "How can I make use of Certawi Certificate Generator?",
+    body: [
+      "Sign up for Certawi for free.",
+      "Pick a template that fits the occasion.",
+      "Choose a colour scheme and upload the necessary information.",
+      "Upload CVS document.",
+      "Choose certificate format (PNG or PDF).",
+      "Generate and download Certificates in Zip file.",
+    ],
+  },
+  {
+    title: "How much does creating a certificate cost?",
+    body: [
+      "It will cost nothing to make custom certificates. Certawi Certificate Generator can be used for free. However, if you want more features, you can upgrade your plan.",
+    ],
+  },
+  {
+    title: "Can I generate Certificates in bulk?",
+    body: [
+      "Definitely! Certawi was created to help you generate digital certificates in bulk.",
+    ],
+  },
+  {
+    title: "How is Certawi different from other Online Certificate Generators?",
+    body: [
+      "Unlike other Certificate Generators, we do not want users to be confused when faced with a wide array of templates that are more or less the same. Instead, Certawi Certificate Generator offers six simple design templates that are distinguishably different from one another and easy for users to use.",
+    ],
+  },
+  {
+    title: "Can I share my account with another user?",
+    body: [
+      "Unfortunately, users cannot share the same account. Each user has their asset allocated under their own personal folder on Certawi Certificate Generator.",
+    ],
+  },
+  {
+    title: "Can I use Certawi on my phone?",
+    body: [
+      "Yes! Certawi is compatible with phone, tablet, laptop and MacBook.",
+    ],
+  },
+];
+
+export const DownlaodsItems = [
+  {
+    title: "What format or size is available for download order?",
+    body: ["The certificate comes in size 8.5 × 11"],
+  },
+  {
+    title: "Do free downloads of Certificate have Certawi watermark?",
+    body: ["No! Certawi offers free downloads without watermark."],
+  },
+  {
+    title: "Can I print one certificate instead of bulk?",
+    body: [
+      "If you do not want to download the certificate in bulk, Certawi provides you the option of downloading each certificate as a stand alone.",
+    ],
+  },
+  {
+    title: " Why is my image blurry?",
+    body: ["This is probably because the image you uploaded is less than 5kb."],
+  },
+];
+
+export const UploadsItems = [
+  {
+    title:
+      "What formats and sizes box CSV file, brand logo and signature is required to be upload?",
+    body: [
+      "To ensure maximum quality download, the CSV file should not be more than 30mb. The required size for the brand logo and signature is 5mb and 25kb respectively.",
+    ],
+  },
+  {
+    title: "What if I made a mistake while uploading my CSV file?",
+    body: [
+      "Certawi Certificate Generator records whatever information you upload. If there is a mistake, you have to re-upload the correct file to the website.",
+    ],
+  },
+
+  {
+    title: "Can I use my Brand Kit with the certificate maker?",
+    body: [
+      "Absolutely. When you have a Brand Kit set up in Certawi Certificate Generator, it is available anytime you want to create certificates. To have a Brand Kit you need to upgrade from a free plan.",
+    ],
+  },
+];
+
+export const FeaturedItems = [
+  {
+    title:
+      "What if I modified my certificate design after I confirmed download?",
+    body: [
+      "Certawi auto saves your designs every 10 seconds to make it easy to use. So, once you modify your design, it automatically saves it. You recent update will be reflected on your design when you generate it.",
+    ],
+  },
+  {
+    title: "Can I make copies of my certificate? ",
+    body: [
+      "Yes! You can create copies on any of your designs. When you edit a copy, it will not make any changes to the original certificate.",
+    ],
+  },
+  {
+    title: "What is the best certificate size?",
+    body: [
+      "The default size Certawi offers is 8.5 × 11. This is also the best certificate size",
+    ],
+  },
+  {
+    title: "Can the size of the certificate be edited?",
+    body: [
+      "Aside from the default size, you can also print the certificate in half the original size or even double.",
+    ],
+  },
+  {
+    title: "What are modifiable in the certificate edit page?",
+    body: "You can edit the colour scheme to suit your preference",
+  },
+];
