+++ frontend/src/pages/FAQ/faq.module.scss
@@ -1,38 +1,37 @@
-@import url("https://fonts.googleapis.com/css2?family=Red+Rose:wght@300;400&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
-
 @mixin flexContainer($direction, $width, $justify) {
   display: flex;
   justify-content: $justify;
   align-items: center;
   flex-direction: $direction;
   width: $width;
+  font-family: "Work Sans", sans-serif;
 }
 
 @mixin container {
   margin: 0 auto;
 }
-* {
-  font-family: "Roboto", sans-serif;
-}
+
 
 .header {
   @include flexContainer(column, 80%, center);
   @include container();
   text-align: center;
+  font-family: "Work Sans", sans-serif;
 
   h2 {
-    font-size: 2rem;
+    font-size: 1.125rem;
     font-weight: 700;
   }
   p {
-    font-size: 1rem;
+    font-size: 0.75rem;
   }
 }
 
 .faqContainer {
   @include container();
   width: 80%;
-  margin-top: 3rem;
+  margin-top: 5rem;
+  font-family: "Work Sans", sans-serif;
 
   h2 {
     font-size: 1.2rem;
@@ -46,29 +45,27 @@
     border-radius: 10px;
     padding: 1rem 2rem;
     margin-bottom: 0.2rem;
+    	font-family: "Work Sans", sans-serif;
 
     .faqQuestions {
       @include flexContainer(row, 100%, space-between);
+      align-items: flex-start;
       gap: 20px;
+      cursor: pointer;
 
       .question {
         display: flex;
-        align-items: center;
+        align-items: flex-start;
         gap: 10px;
+        
         h3 {
-          line-height: 1.5;
+          line-height: 160%;
           font-weight: 500;
-          font-size: 0.75rem;
+          font-size:0.875rem;
           text-align: left;
         }
 
-        .faqIcon {
-          width: 50px;
-          height: 50px;
-          img {
-            width: 100%;
-          }
-        }
+       
       }
       .icon {
         font-size: 2.5rem;
@@ -78,6 +75,7 @@
       width: 80%;
       margin: 0 auto;
       color: #656565;
+      transition: all .3s cubic-bezier(0.445, 0.05, 0.55, 0.95);
       li {
         line-height: 1.5;
         font-size: 0.9rem;
@@ -95,12 +93,9 @@
     // text-align: center;
     // font-family: "Roboto", sans-serif;
 
-    h2 {
-      font-size: 1.5rem;
-      font-weight: 600;
-    }
+  
     p {
-      font-size: 1rem;
+    
       font-weight: 400;
       padding: 0.5rem 5rem;
     }
@@ -126,7 +121,7 @@
 
         .question {
           display: flex;
-          align-items: center;
+          align-items: flex-start;
           gap: 10px;
           h3 {
             line-height: 1.5;
@@ -167,7 +162,7 @@
   align-items: center;
 }
 
-@media screen and (min-width: 769px) {
+@media screen and (min-width: 768px) {
   .header {
     max-width: 1440px;
     // @include flexContainer(column, 70%);
@@ -176,11 +171,11 @@
     // font-family: "Roboto", sans-serif;
 
     h2 {
-      font-size: 2.5rem;
-      font-weight: 600;
+      font-size: 2.25rem;
+      
     }
     p {
-      font-size: 1.2rem;
+      font-size: 1.25rem;
       font-weight: 400;
       padding: 0.5rem 5rem;
     }
@@ -216,9 +211,7 @@
           .faqIcon {
             width: 50px;
             height: 50px;
-            img {
-              width: 100%;
-            }
+           
           }
         }
         .icon {
